+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraVideoChannel.kt
@@ -106,13 +106,6 @@ class SoraVideoChannel(
             disconnect()
         }
 
-        override fun onError(mediaChannel: SoraMediaChannel, reason: SoraErrorReason) {
-            SoraLogger.d(TAG, "[video_channel] @onError $reason")
-            handler.post {
-                listener?.onError(this@SoraVideoChannel, reason)
-            }
-        }
-
         override fun onError(mediaChannel: SoraMediaChannel, reason: SoraErrorReason, message: String) {
             SoraLogger.d(TAG, "[video_channel] @onError $reason: $message")
             handler.post {